Move buid folder to /docs for gh pages

+++ src/main.rs
@@ -168,13 +168,13 @@ fn render(app: &App) -> Result<(), JsValue> {
                 i,
                 j,
                 j,
-                app.choices[i][j].times_left as f64 / app.total_paths,
+                app.choices[i][j].times_left as f64 / app.total_paths as f64,
             )?;
             app.draw_segment(
                 i,
                 j,
                 j + 1,
-                app.choices[i][j].times_right as f64 / app.total_paths,
+                app.choices[i][j].times_right as f64 / app.total_paths as f64,
             )?;
         }
